% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoDistSampCovar.R
\name{autoDistSampCovar}
\alias{autoDistSampCovar}
\title{Fit multiple distance sampling detection functions with different covariates}
\usage{
autoDistSampCovar(detectionData, siteData, covars, likelihoods = c("halfnorm",
  "hazrate", "uniform", "negexp", "Gamma"), series = c("cosine", "hermite",
  "simple"), expansions = 0:3, w.lo = 0, w.hi = NULL,
  pointSurvey = FALSE, warn = TRUE, area = 1, ci = 0.95, R = 500,
  bySite = FALSE, includeFits = FALSE, plot.bs = FALSE,
  showProgress = TRUE, plot = TRUE, criterion = "AICc", ...)
}
\arguments{
\item{detectionData}{This parameter is passed to \code{dfuncEstim} 
and \code{abundEstim}. See \code{abundEstim} documentation for definition.}

\item{siteData}{This parameter is passed to \code{abundEstim}.
See \code{abundEstim} documentation for definition.}

\item{covars}{Character vector, names of cols in detectionData to use as covariates.}

\item{likelihoods}{Vector of strings specifying the 
likelihoods to consider during model selection. Valid values 
at present are "uniform", "halfnorm", "hazrate", "negexp", 
and "Gamma". See Details for the models this routine considers.}

\item{series}{Vector of series types to consider during model selection. 
Valid values are 'simple', 'hermite', and 'cosine'. See Details for 
the models this routine considers.}

\item{expansions}{Vector of the number of expansion terms to 
consider during model selection. Valid values are 0 through 3. 
See Details for the models this routine considers. Note, expansion terms are not currently
allowed in models with covariates.}

\item{w.lo}{This parameter is passed to \code{dfuncEstim}.
See \code{dfuncEstim} documentation for definition.}

\item{w.hi}{This parameter is passed to \code{dfuncEstim}.
See \code{dfuncEstim} documentation for definition.}

\item{pointSurvey}{This parameter is passed to \code{dfuncEstim}.
See \code{dfuncEstim} documentation for definition.}

\item{warn}{This parameter is passed to \code{dfuncEstim}.
\code{dfuncEstim} documentation for definition.}

\item{area}{This parameter is passed to \code{abundEstim}.
See \code{abundEstim} documentation for definition.}

\item{ci}{This parameter is passed to \code{abundEstim}.
See \code{abundEstim} documentation for definition.}

\item{R}{This parameter is passed to \code{abundEstim}.
See \code{abundEstim} documentation for definition.}

\item{bySite}{This parameter is passed to \code{abundEstim}.
See \code{abundEstim} documentation for definition.}

\item{plot.bs}{Logical for whether to plot bootstrap 
iterations after the top model has been selected and 
during final estimation of confidence intervals.  
This parameter is passed unchanged to \code{abundEstim}.
See \code{abundEstim} help for additional information.}

\item{showProgress}{Logical for whether to 
suppress intermediate output.  If \code{showProgress=TRUE}, 
a table of model fitting results appears in the console as they 
are estimated, and a progress bar shows progress through 
the bootstrap iterations at the end.  If \code{showProgress=FALSE}, 
all intermediate output is suppressed which is handy for programming
and simulations.}

\item{plot}{Logical scalar specifying whether to plot models during model selection. 
If \code{TRUE}, a histogram with fitted distance function is plotted for every fitted model. 
The function pauses between each plot and prompts the user for whether they want to continue or not. 
For completely automated estimation, set \code{plot} = \code{FALSE}.}

\item{criterion}{A string specifying the criterion to use when assessing model fit.
The best fitting model from this routine is the one with lowest value 
of this fit criterion.  This must be one of "AICc" (the default), 
"AIC", or "BIC".  See \code{\link{AIC.dfunc}} for formulas.}

\item{...}{Additional parameters passed to \code{dfuncEstim}, which in turn are passed to \code{F.gx.estim}. 
These include \code{x.scl}, \code{g.x.scl}, and \code{observer} for estimating double observer probabilities.}

\item{covars}{Character vector, names of cols in detectionData to use as covariates.}
}
\value{
A list with R+2 elements where R is the number of models run.
Elements 1:R are the dfunc objects fit using Rdistance, element R+1 is the data.frame of the AICc table,
and element R+2 is the detectionData used in the detection functions (which had truncation applied, including the
re-scaling of distances if w.lo > 0).
}
\description{
Runs multiple combinations of likelihoods and covariates, and creates an AICc table
}
\details{
During model selection, each series and number of expansions is crossed with 
each of the likelihoods. For example, if \code{likelihoods} has 3 elements, 
\code{series} has 2 elements, and \code{expansions} has 4 elements, 
the total number of models fitted is 3 (likelihoods) * 2 (series) * 4 (expansions) 
= 24 models.  The default specification fits 41 detection functions 
from the "halfnorm", "hazrate", "uniform", "negexp", and "Gamma" likelihoods 
(note that Gamma does not currently implement expansions, see 
\code{\link{Gamma.like}}). Note, expansion terms are not currently
allowed in models with covariates.  The model with lowest AIC is selected 
as 'best', and estimation of abundance proceeds using that model.

Suppress all intermediate output using \code{plot.bs=FALSE}, 
\code{showProgress=FALSE}, and \code{plot=FALSE}.
}
\examples{
# Automate fitting detection functions for different comninations of likelihoods and covariates
# Load example sparrow data (line transect survey type)
data(sparrowDetectionData)
data(sparrowSiteData)

ms<- dfuncModSel(detectionData = sparrowDetectionData, siteData = sparrowSiteData,  covars=c("1", "bare", "shrub", "herb", "height",
 "bare + observer", "shrub + observer", "herb + observer", "height + observer"),likelihoods=c("halfnorm", "hazrate"), 
 w.lo=0, w.hi=125)
}
\seealso{
\code{\link{dfuncEstim}}, \code{\link{abundEstim}}.
}
\author{
Trent McDonald, WEST Inc.,  \email{tmcdonald@west-inc.com}\cr
        Aidan McDonald, WEST Inc.,  \email{aidan@mcdcentral.org}\cr
        Jason Carlisle, University of Wyoming and WEST Inc., \email{jcarlisle@west-inc.com}\cr
        Abigail Hoffman, Univeristy of Wyoming and WEST Inc., \email{ahoffm12@uwyo.edu}
}
\keyword{model}
