% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcCI.R
\name{bcCI}
\alias{bcCI}
\title{Bias corrected bootstraps}
\usage{
bcCI(x.bs, x, ci = 0.95)
}
\arguments{
\item{x.bs}{A vector of bootstrap estimates of some quantity.}

\item{x}{A scalar of the original estimate of the quantity.}

\item{ci}{A scalar of the desired confidence interval
coverage.}
}
\value{
A named vector containing the lower and upper endpoints of
the bias-corrected bootstrap confidence interval.
}
\description{
Calculate bias-corrected confidence intervals for bootstrap data
using methods in Manly textbook.
}
