% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUnits.R
\name{checkUnits}
\alias{checkUnits}
\title{Check for the presence of units}
\usage{
checkUnits(ml)
}
\arguments{
\item{ml}{An \code{Rdistance} model list produced by 
\code{\link{parseModel}} containing a list
of parameters for the distance model.}
}
\value{
The input \code{ml} list, with units of various 
quantities converted to common units.  If a check fails, for example, 
a quantity does not have units, an error is thrown.
}
\description{
Check for the presence of physical measurement units on key
columns of  
an \code{RdistDf} data frame.
}
