% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective.radius.R
\name{effective.radius}
\alias{effective.radius}
\title{Effective Transect Radius for estimated detection functions with point transects}
\usage{
effective.radius(obj)
}
\arguments{
\item{obj}{An estimated detection function object.  An estimated detection
function object has class 'dfunc', and is usually produced by a call to 
\code{F.dfunc.estim}. The estimated detection function may optionally contain 
a \eqn{g(0)} component.  If no \eqn{g(0)} component is found, \eqn{g(0)} = 1
is assumed.}
}
\value{
A scalar equal to the area under the detection function from \code{obj$w.lo} to \code{obj$w.hi}.
}
\description{
Computes effective transect radius for estimated detection functions with point transects.
}
\details{
The equivalent of Effective Strip Width for point transects.
}
\seealso{
\code{\link{F.dfunc.estim}} \code{\link{ESW}}
}
\author{
Aidan McDonald, WEST, Inc., \email{aidan@mcdcentral.org}
}
\keyword{modeling}
