% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaModes.R
\name{GammaModes}
\alias{GammaModes}
\title{Modes of the Gamma distribution}
\usage{
GammaModes(params)
}
\arguments{
\item{params}{A matrix of Gamma distribution parameters. First column 
is the scale parameter, and is a fuction of covariates.
Second column is the shape parameter.}
}
\value{
A vector of the locations of the gamma modes associated 
with each row in the \code{params} matrix.
}
\description{
Compute mode (i.e., maximum) of Rdistance's version of the gamma 
distribution.
}
