% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huber.cumFunc.R
\name{huber.cumFunc}
\alias{huber.cumFunc}
\title{Huber Cumulative Function}
\usage{
huber.cumFunc(x, t1, t2, p, w)
}
\arguments{
\item{x}{A vector of distances. Can have units or not
(i.e., regular numeric).}

\item{t1}{A vector of values for the \eqn{\theta_1}{T1} parameter 
of the `huber` likelihood. If \code{x} has units, \code{t1} must have 
compatible units.}

\item{t2}{A vector of values for the \eqn{\theta_2}{T2} parameter 
of the `huber` likelihood. If \code{x} has units, \code{t2} must have 
compatible units.}

\item{p}{A vector of values for the \eqn{p} parameter of the 
`huber` likelihood. If \code{x} has units, \code{p} must have 
units of '[1]' (i.e., \code{setUnits(p,1)}).}

\item{w}{A vector of maximum strip widths, the maximum distance. 
If \code{x} has units, \code{x} must have 
compatible units.}
}
\value{
A vector of values from the `huber` cumulative function. 
The `huber` cumulative function is
\deqn{
F(x|\theta_1,\theta_2,p) = \int_0^x f(y|\theta_1,\theta_2,p) dy,}{
F(x|T1,T2,p) = Area Under f(y|T1,T2,p) from 0 to x,
}
where \eqn{f(y|\theta_1,\theta_2,p)}{f(y|T1,T2,p)} is Rdistance's `huber` 
likelihood. The only difference between this \emph{cumulative function}, 
and the \emph{cumulative distribution function}
is the scaling constant. That is, the maximum of the \emph{cumulative 
function} is greater than 1 while the maximum \emph{cumulative distribution function}
is exactly 1.
}
\description{
Computes the cumulative function of the `huber` likelihood. The 
cumulative function is proportional to the `huber` 
cumulative distribution function, differing only by appropriate
scaling constant.
}
\examples{
d <- -10:210

# Cumulative function
fd <- huber.cumFunc(d, 125, 25, .05, 200) 
plot(d, fd, type="l")

# Cumulative distribution function
Fd <- fd / huber.cumFunc(200, 125, 25, .05, 200)

}
\seealso{
\code{\link{huber.like}}
}
