% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertOneStepBreaks.R
\name{insertOneStepBreaks}
\alias{insertOneStepBreaks}
\title{Insert oneStep Likelihood breaks}
\usage{
insertOneStepBreaks(obj, newData, xseq)
}
\arguments{
\item{obj}{A fitted Rdistance model object}

\item{newData}{A data frame containing covariate values to 
use in prediction.}

\item{xseq}{A vector of distances into which the break points 
are inserted.}
}
\value{
A vector like \code{xseq}, but with the break points 
inserted.
}
\description{
Compute break points in a onestep likelihood
and insert them into a sequence of distances.  The idea 
is to insert a point just left and just right of the breaks 
so that they plot as vertical lines.
}
