% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateNegexp.R
\name{integrateNegexp}
\alias{integrateNegexp}
\title{Integrate Negative exponential}
\usage{
integrateNegexp(object, newdata = NULL)
}
\arguments{
\item{object}{An Rdistance model frame or fitted distance function,
normally produced by a call to \code{\link{dfuncEstim}}.}

\item{newdata}{A data frame containing new values for 
covariates at which either
ESW's or EDR's will be computed. If NULL and 
  \code{object} contains covariates, the  
  covariates stored in
  \code{object} are used (like \code{\link{predict.lm}}).
  If not NULL, covariate values in \code{newdata}
  are used. 
  See \bold{Value} section for more information.}
}
\value{
A vector of areas under distance functions. 
If \code{newdata} is specified, return length is 
\code{nrow(newdata)}.  If \code{newdata} is NULL, 
return length is \code{length(distances(object))}.
}
\description{
Compute integral of the negative exponential distance function.
}
\examples{

# Fake distance function object w/ minimum inputs for integration
d <- units::set_units(rep(1,4),"m") # Only units needed, not values
obs <- factor(rep(c("obs1", "obs2"), 2))
beta <- c(-5, -0.5)
w.hi <- 125
w.lo <- 20
ml <- list(
    mf = model.frame(d ~ obs)
  , par = beta 
  , likelihood = "negexp"
  , w.lo = units::set_units(w.lo, "m")
  , w.hi = units::set_units(w.hi, "m")
)
class(ml) <- "dfunc"
integrateNegexp(ml)

# Check: Integral of exp(-bx) from 0 to w.hi-w.lo
b <- c(exp(beta[1]), exp(beta[1] + beta[2]))
intgral <- (1 - exp(-b*(w.hi - w.lo))) / b
intgral


}
\seealso{
\code{\link{integrateNumeric}}; \code{\link{integrateNegexp}}; 
\code{\link{integrateOneStep}}
}
