% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isUnitless.R
\name{isUnitless}
\alias{isUnitless}
\title{isUnitless - Test whether object is unitless}
\usage{
isUnitless(obj)
}
\arguments{
\item{obj}{A numeric scaler or vector, with or without units.}
}
\value{
TRUE if \code{obj} has units and they 
are either '[1]' or the denominator units equal
the numerator units.  Otherwise, return FALSE.
If \code{obj} does not have units, this routine
returns TRUE.
}
\description{
Tests whether a 'units' object is 
unitless.  If something is unitless, sometimes it is 
better to remove it's units before calculations. 
Unitless objects such as ratios sometimes get assigned
units of '[1]', which is preferable, and 
sometimes they have units like '[m/m]'.  The 
\code{units} package should convert '[m/m]' to 
'[1]', but it does not always.
}
\examples{
a <- units::set_units(2, "m")
b <- a / a
isUnitless(a)
isUnitless(b)
isUnitless(3)

}
