% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.like.R
\name{logistic.like}
\alias{logistic.like}
\title{logistic.like - Logistic distance function likelihood}
\usage{
logistic.like(a, dist, covars)
}
\arguments{
\item{a}{A vector or matrix of covariate 
and expansion term 
coefficients. Dimension is $k$ X $p$, where 
$k$ (i.e., \code{nrow(a)}) is the number of coefficient
vectors to evaluate (cases) and $p$ (i.e., \code{ncol(a)})
is the number of covariate and expansion 
coefficients in the likelihood. If \code{a} is a 
dimensionless vector, it is interpreted to be a 
single row with $k$ = 1. 
Covariate coefficients in \code{a} are the first 
$q$ values ($q$ <= $p$), and must be on a log scale.}

\item{dist}{A numeric vector of length $n$ or 
a single-column matrix (dimension $n$X1) containing 
detection distances at which to evaluate the likelihood.}

\item{covars}{A numeric vector of length $q$ or 
matrix of dimension $n$X$q$ containing covariate values 
associated with distances in argument \code{d}}
}
\value{
A list containing the following two components:
\itemize{
  \item \code{\bold{L.unscaled}}: A matrix of size $n$X$k$X$b$ 
   containing likelihood values evaluated at 
   distances in \code{dist}.
    Each row is associated with 
   a single distance, and each column is associated with 
   a single case (row of \code{a}).  This matrix is  
   "unscaled" because the underlying likelihood does 
   not integrate to one. Values in \code{L.unscaled} 
   are always greater than or equal to zero.
   
 \item \code{\bold{params}}: A $n$X$k$X$b$ array of the 
 likelihood's (canonical) parameters, First page contains 
 parameter values related to covariates (i.e., $s = exp(x'a)$),
 while subsequent pages contain other parameters. 
 $b$ = 1 for halfnorm, negexp; $b$ = 2 for hazrate and 
 others.
 Rows correspond to distances in \code{dist}. Columns 
 correspond to rows from argument \code{a}. 
}
}
\description{
Computes a two parameter logistic distance function.
}
\details{
The 'logistic' likelihood contains two 
parameters.  Parameter \eqn{a} determines the scale and is 
labeled 'threshold' in Rdistance.  Parameter \eqn{b} determines 
sharpness (slope) of the likelihood's decrease at \eqn{a} and is labeled
'knee' in Rdistance.  
(This function is called the 
\emph{heavy side} function in engineering).  The form 
of the function is, 
\deqn{f(x|a,b) = 1 - \frac{1}{1 + \exp(-b(x-a))} = 
\frac{\exp( -b(x-a) )}{1 + exp( -b(x-a) )},}{%
f(x|a,b) = 1 - 1 / (1 + exp(-b*(x-a))) = exp(-b*(x-a)) / (1 + exp(-b*(x-a))).} 

Parameter \eqn{a} is the location (distance) where the 
(unscaled) likelihood equals  
0.5. That is, the inverse likelihood of 0.5 
is \eqn{a}, before scaling; or, 
\code{logistic.like( c(a,b), a, matrix(1,1,1))} equals 
\code{0.5}). 

Parameter \eqn{b} is slope of function 
at \eqn{a}.  
Prior to scaling, 
slope of the likelihood at \eqn{a} is \eqn{-b/4}. 
If \eqn{b}
is large, the "knee" is sharp and the likelihood can look 
uniform with support from 
\code{w.lo} to \eqn{a/f(0)}.  If \eqn{b} is small, the 
"knee" is shallow and the density of observations declines 
in an elongated "S" shape pivoting at \eqn{a/f(0)}.  
As \eqn{b} grows large and assuming f(0) = 1, the effective 
strip width approaches \eqn{a}.  

See plots in Examples.
}
\examples{
d <- seq(0, 100, length=100)
covs <- matrix(1,length(d),1)
logistic.like(c(log(20), 1), d, covs)

# Changing threshold parameter
plot(x, logistic.like(c(log(20), 20), x, covs)$L.unscaled, type="l", col="red")
lines(x, logistic.like(c(log(40), 20), x, covs)$L.unscaled, col="blue")
ablines(h = 0.5, lty = 2)
ablines(v = c(20,40), lty = 2)

# Changing knee parameter
plot(x, logistic.like(c(log(50), 100), x, covs)$L.unscaled, type="l", col="red")
lines(x, logistic.like(c(log(50), 1), x, covs)$L.unscaled, col="blue")
lines(x, logistic.like(c(log(50), .1), x, covs)$L.unscaled, col="orange")

         
}
\seealso{
\code{\link{dfuncEstim}},
         \code{\link{hazrate.like}},
         \code{\link{uniform.like}},
         \code{\link{negexp.like}},
         \code{\link{Gamma.like}},
         \code{\link{triangle.like}}
}
