% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlEstimates.R
\name{mlEstimates}
\alias{mlEstimates}
\title{Distance function maximum likelihood estimates}
\usage{
mlEstimates(ml, strt.lims)
}
\arguments{
\item{ml}{Either a Rdistance 'model frame' or an Rdistance 
'fitted object'.  Both are of class "dfunc". 
Rdistance 'model frames' are lists containing components 
necessary to estimate a distance function, but no estimates. 
Rdistance 'model frames' are typically
produced by calls to \code{\link{parseModel}}. 
Rdistance 'fitted objects'
are typically produced by calls to \code{\link{dfuncEstim}}.
'Fitted objects' are 'model frames'
with additional components such as the parameters estimates, 
log likelihood value, convergence information, and the variance-
covariance matrix of the parameters.}

\item{strt.lims}{A list containing start, low, and high limits for 
parameters of the requested likelihood. This list is typically produced
by a call to \code{\link{startLimits}}.}
}
\value{
An Rdistance fitted model object. This object contains the 
raw object returned by the optimization routine (e.g., \code{nlming}), 
and additional components specific to Rdistance.
}
\description{
Estimate parameters of a distance function using maximum likelihood.
}
