% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.dfunc.R
\name{model.matrix.dfunc}
\alias{model.matrix.dfunc}
\title{Rdistance model matrix}
\usage{
\method{model.matrix}{dfunc}(object, \dots)
}
\arguments{
\item{object}{An Rdistance model frame or fitted distance function,
normally produced by a call to \code{\link{dfuncEstim}}.}

\item{...}{Ignored}
}
\value{
A matrix containing covariates for fitting an
Rdistance model.
}
\description{
Extract the model matrix ("X" matrix) from an Rdistance model object.
}
\examples{

data(sparrowDf)
sparrowModel <- parseModel( sparrowDf, dist ~ observer )
model.matrix(sparrowModel)

}
