% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscarsLimits.R
\name{oscarsLimits}
\alias{oscarsLimits}
\title{Modify start and limits for OSCARS optimization}
\usage{
oscarsLimits(origLims, ml)
}
\arguments{
\item{origLims}{A list of $start, $low, and $high components
that are vectors of the same length.  $start are the starting
values, $low are the lower bounds, and $high are the upper bounds
on the coefficients in the model.}

\item{ml}{A "model list" containing the problem parameters.}
}
\value{
A list just like \code{curLims} with improved start and
limits.  In addition, the return contains a couple quantities used
later to assess convergence (i.e., $startConvergence and
$startEvaluations)
}
\description{
Improve (tighten) the starting values and
limits for the OSCARS algorithm. OSCARS needs tighter bounds
because of it's global nature.
}
\details{
Despite several likelihoods being discontinuous, gradient based
methods do amazingly well.  We improve (tighten) the
uninformed start limits by estimating the model
using Newton-Raphson and constructing limits from those estimates.
}
