% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdistance-package.R
\docType{package}
\name{Rdistance-package}
\alias{Rdistance-package}
\alias{distance}
\alias{Rdistance}
\alias{point-transect}
\alias{line-transect}
\title{Rdistance - Distance Sampling Analyses for Abundance Estimation}
\description{
\code{Rdistance} contains functions and associated routines to analyze
distance-sampling data collected on point or line transects.
Some of \code{Rdistance}'s features include:
\itemize{
\item Accommodation of both point and line transect analyses in one routine (\code{\link[=dfuncEstim]{dfuncEstim()}}).
\item Regression-like formula for inclusion of distance function covariates (\code{\link[=dfuncEstim]{dfuncEstim()}}).
\item Automatic bootstrap confidence intervals (\code{\link[=abundEstim]{abundEstim()}}).
\item Parallel processing of bootstrap iterations (\code{parallel} argument of \code{\link[=abundEstim]{abundEstim()}}).
\item Availability of both study-area and site-level abundance estimates (\code{help("predict.dfunc")}).
\item Rigorous physical measurement requirements and automated conversion when necessary (\code{\link{\%#\%}}, \code{\link[=setUnits]{setUnits()}}).
\item Classic parametric distance functions (\code{\link[=halfnorm.like]{halfnorm.like()}}, \code{\link[=hazrate.like]{hazrate.like()}}, \code{\link[=negexp.like]{negexp.like()}}), and
expansion functions (\code{\link[=cosine.expansion]{cosine.expansion()}}, \code{\link[=hermite.expansion]{hermite.expansion()}}, \code{\link[=simple.expansion]{simple.expansion()}}).
\item Mixture distance functions for non-standard shapes and thresholds (\code{\link[=oneStep.like]{oneStep.like()}}, \code{\link[=triangle.like]{triangle.like()}}, and \code{\link[=huber.like]{huber.like()}}).
\item Automated distance function fitting and selection \code{\link[=autoDistSamp]{autoDistSamp()}}.
\item \code{print}, \code{plot}, \code{predict}, \code{coef}, and \code{summary} methods for distance function objects and
abundance classes.
}
}
\section{Background:}{
Distance-sampling is a popular method for abundance estimation in
ecology. Line transect surveys are conducted by traversing
randomly placed transects in a study area with the objective of
sighting animals and estimating density or abundance.  Data collected
during line transect surveys consists of \emph{target} sightings,
either of individuals or groups, and off-transect distances to the
original location of the \emph{target}.
When \emph{targets} are sighted in groups, data include the number
of individuals in the group.

Point transect surveys are similar except that observers stop one
or more times along the transect to observe \emph{targets}.  Point transects are
popular avian survey methods where detections are often auditory
cues. Point transects are also for studies using automated auditory detectors
or trail cameras.  Point transect data consists of radial distances from
the observer to the \emph{target}.

The defining feature of distance sampling is the tendency for
probability of detection to decline as
off-transect or radial distances increase. Targets far from
the observer are generally
harder to detect than those closer.  In most
line transect studies, \emph{targets} on the transect (off-transect
distance = 0) are assumed to be sighted with 100\% probability.  This
assumption allows researchers to estimate the proportion of missed
targets and in turn adjust the number of
sighted \emph{targets} for missed detections.
Some studies utilize two observers searching the same areas and are able to
estimate the proportion of individuals missed on the transect line and
thereby eliminate the
assumption that all individuals on the line have been observed.
}

\section{Purpose:}{
The author's aims are
to provide an easy-to-use, rigorous, and flexible analysis option in R for
distance-sampling data.  The authors believe that beginning
users need easy-to-use and easy-to-understand software,
while advanced users require greater flexibility and customization, and
their aim is to meet the demands of both groups.
}

\section{Data sets:}{
\code{Rdistance} contains the following example data sets:
\itemize{
\item Line-transect sampling of Brewers sparrows in central Wyoming
(\code{\link[=sparrowDf]{sparrowDf()}}).
\item Point-transect sampling of Sage Thrashers in central Wyoming
(\code{\link[=thrasherDf]{thrasherDf()}}).
}
}

\section{References}{
Buckland, S.T., Anderson, D.R., Burnham, K.P. and Laake, J.L.
1993. \emph{Distance Sampling: Estimating Abundance of Biological
Populations}. Chapman and Hall, London.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://mcdonalddatasciences.com/Rdistance.html}
  \item Report bugs at \url{https://github.com/tmcd82070/Rdistance/issues}
}

}
\author{
Main author and maintainer: Trent McDonald
\href{mailto:trent@mcdonalddatasciences.com}{trent@mcdonalddatasciences.com}

Coauthors: Ryan Nielson, Jason Carlisle, and Aidan McDonald

Contributors: Ben Augustine, James Griswald, Joel Reynolds, Pham Quang, Earl
Becker, Aaron Christ, Brook Russelland, Patrick McKann, Lacey Jeroue,
Abigail Hoffman, Michael Kleinsasser, and Ried Olson
}
\keyword{package}
