% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smlAntelopeDetections.R
\docType{data}
\name{smlAntelopeDetections}
\alias{smlAntelopeDetections}
\title{Small antelope detection data on walking transects}
\format{
A data.frame containing 129 rows and 11 columns.  Each row represents
a detected group of antelope.  Column descriptions: 
\enumerate{ 
  \item \code{siteID}: ID for each row.  Row ID's consist of 'year_season_transect.  
  \item \code{transect_bearing}: Compass bearing of the transect 
  (direction of travel) at the moment of detection. 
  \item \code{obs_bearing}: Compass bearing of the antelope group (at their 
  initial location) from the searcher's location at the moment of detection.
  \item \code{obs_dist}: Distance from searcher to antelope group's location, 
  at the moment of detection. 
  \item \code{number}: Number of antelope in the sighted group.
  \item \code{perpDist}: Perpendicular off-transect distance assuming searchers 
  continued straight from the moment of detection. 
  \item \code{species}: Species of antelope detected. 
}
}
\source{
The small antelope data is a subset of walking transect data collected 
by Lion Landscapes (\url{https://www.lionlandscapes.org}).
}
\description{
A data set containing detections of small antelope (Grey Duiker, Grysbok, and Klipspringer)
on a study area in central Africa. The overall study area was partitioned into 
3km X 3km grid cells. Trackers and guides walked one north-bound 3km transect 
1 km from, and parallel to, the western boundary of each cell. 
Trackers and guides also walked a south-bound 3km transect 1 km from, and parallel
to, the eastern boundary of each cell. Transects were north-south 
in overall direction, but searchers deviated around obstacles (vegetation, impassable 
terrain) as necessary.  Size of the study area was defined to be coverage 
of all 9 km^2 grid cells.

During surveys, trackers and guides recorded all species seen; but, this data 
set contains only observations of small antelope. 
Searchers recorded distance from searcher location to original sighting 
location (\code{obs_dist}), compass bearing of the sighting location 
(\code{obs_bearing}), compass bearing of the transect (\code{transect_bearing}), 
and number of antelope seen in the group (\code{number}).  
Analysts computed perpendicular off-transect distances (\code{perpDist}) using  
bearing differences and trigonometry (i.e., x = r / sin(theta)). 

Initially, analysts combined sightings of all three small antelope species 
for estimation of a detection function, but estimated abundance separately  
for each species during each survey. The data set contains information 
from 7 surveys between early 2016 and early 2022 on 35 grid cells.  During 
most surveys, all grid cells were surveyed (35 cells * 2 transects = 
70 transects); but, during the last three surveys 1, 2, and 5 grid cells, 
respectively, could not be surveyed due to safety concerns.
}
\seealso{
\code{\link{smlAntelopeTransects}}
}
\keyword{datasets}
