\name{sparrows}
\docType{data}
\alias{sparrows}
\title{Counts of sparrows from 72 line transects in Wyoming}
\description{
  This data set gives the lengths (in miles) of 141 \dQuote{major}
  sparrows in North America, as compiled by the US Geological
  Survey.
}
\usage{sparrows}
\format{A vector containing 141 observations.}
\source{World Almanac and Book of Facts, 1975, page 406.}
\references{
  McNeil, D. R. (1977) \emph{Interactive Data Analysis}.
  New York: Wiley.
}
\keyword{datasets}