% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform.like.R
\name{uniform.like}
\alias{uniform.like}
\title{uniform.like - Uniform distance likelihood}
\usage{
uniform.like(a, dist, covars)
}
\arguments{
\item{a}{A vector or matrix of covariate 
and expansion term 
coefficients. Dimension is $k$ X $p$, where 
$k$ (i.e., \code{nrow(a)}) is the number of coefficient
vectors to evaluate (cases) and $p$ (i.e., \code{ncol(a)})
is the number of covariate and expansion 
coefficients in the likelihood. If \code{a} is a 
dimensionless vector, it is interpreted to be a 
single row with $k$ = 1. 
Covariate coefficients in \code{a} are the first 
$q$ values ($q$ <= $p$), and must be on a log scale.}

\item{dist}{A numeric vector of length $n$ or 
a single-column matrix (dimension $n$X1) containing 
detection distances at which to evaluate the likelihood.}

\item{covars}{A numeric vector of length $q$ or 
matrix of dimension $n$X$q$ containing covariate values 
associated with distances in argument \code{d}}
}
\value{
A list containing the following two components:
\itemize{
  \item \code{\bold{L.unscaled}}: A matrix of size $n$X$k$X$b$ 
   containing likelihood values evaluated at 
   distances in \code{dist}.
    Each row is associated with 
   a single distance, and each column is associated with 
   a single case (row of \code{a}).  This matrix is  
   "unscaled" because the underlying likelihood does 
   not integrate to one. Values in \code{L.unscaled} 
   are always greater than or equal to zero.
   
 \item \code{\bold{params}}: A $n$X$k$X$b$ array of the 
 likelihood's (canonical) parameters, First page contains 
 parameter values related to covariates (i.e., $s = exp(x'a)$),
 while subsequent pages contain other parameters. 
 $b$ = 1 for halfnorm, negexp; $b$ = 2 for hazrate and 
 others.
 Rows correspond to distances in \code{dist}. Columns 
 correspond to rows from argument \code{a}. 
}
}
\description{
Compute the uniform  
distance functions.
}
\details{
The uniform (or constant) likelihood 
is, 
\deqn{f(d|a) = 1}
for \eqn{0 \leq d \leq a}{0 <= d <= a}, and 0 for all
\eqn{d}{d} less than 0 or greater than \eqn{a}{a}.  
Covariates effect parameter \eqn{a}{a} 
via the log link function, i.e., \eqn{a = exp(x'b)},
where \eqn{x} is the vector of covariate values 
associated with distance \eqn{d} and \eqn{b}
is the vector of estimated coefficients. Area under
the likelihood is \eqn{a}.
}
\examples{

d <- seq(0, 100, length=100)
covs <- matrix(1,length(d),1)
uniform.like( log(80), d, covs )

plot(d, uniform.like( log(80), d, covs)$L.unscaled, type = "l", col = "blue")

# Likelihood profile for sparrow data
sparrowDf <- RdistDf(sparrowSiteData, sparrowDetectionData)
dist <- tidyr::unnest(sparrowDf, cols = detections)$dist
covs <- matrix(1,length(dist),1)
a <- seq(75, 210, length= 100)
L <- rep(NA, length(a))
for(i in 1:length(a)){
  y <- uniform.like(log(a[i]), dist, covs)$L.unscaled / a[i]
  y[ !is.na(y) & y <= 0] <- .Machine$double.eps
  L[i] <- -sum(log(y), na.rm = TRUE) 
}
plot(a,L)
abline(v = 146)

}
\seealso{
\code{\link{dfuncEstim}},
         \code{\link{hazrate.like}},
         \code{\link{uniform.like}},
         \code{\link{negexp.like}},
         \code{\link{Gamma.like}},
         \code{\link{triangle.like}}
}
