% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundEstim.R
\name{abundEstim}
\alias{abundEstim}
\title{Estimate abundance from distance-sampling data}
\usage{
abundEstim(
  dfunc,
  detectionData,
  siteData,
  area = NULL,
  ci = 0.95,
  R = 500,
  plot.bs = FALSE,
  bySite = FALSE,
  showProgress = TRUE,
  control = RdistanceControls()
)
}
\arguments{
\item{dfunc}{An estimated 'dfunc' object produced by \code{dfuncEstim}.}

\item{detectionData}{A data frame containing detection distances 
(either perpendicular for line-transect or radial for point-transect
designs), with one row per detected object or group.   
This data frame must contain at least the following 
information: 
\itemize{
  \item Detection Distances: A single column containing 
  detection distances must be specified on the left-hand 
  side of \code{formula}.  As of Rdistance version 2.2.0, 
  the detection distances must have measurement units attached. 
  Attach measurements units to distances using \code{library(units);units()<-}.
  For example, \code{library(units)} followed by \code{units(df$dist) <- "m"} or 
  \code{units(df$dist) <- "ft"} will work. Alternatively, 
  \code{df$dist <- units::set_units(df$dist, "m")} also works.
  
  \item Site IDs: The ID of the transect or point 
  (i.e., the 'site') where each object or group was detected.
  The site ID  column(s) (see arguments \code{transectID} and
  \code{pointID}) must 
  specify the site (transect or point) so that this 
  data frame can be merged with \code{siteData}.    
 
  \item In a later release, \code{Rdistance} will allow detection-level 
  covariates.  When that happens, detection-level 
  covariates will appear in this data frame. 
   
}
See example data set \code{\link{sparrowDetectionData}}.
See also \bold{Input data frames} below 
for information on when \code{detectionData} and 
\code{siteData} are required inputs.}

\item{siteData}{A data.frame containing site (transect or point)
 IDs and any 
\emph{site level} covariates to include in the detection function. 
Every unique surveyed site (transect or point) is represented on
one row of this data set, whether or not targets were sighted 
at the site.  See arguments \code{transectID} and 
\code{pointID} for an explanation of site and transect ID's. 

If sites are transects, 
this data frame must also contain transect length. By 
default, transect length is assumed to be in column 'length' 
but can be specified using argument \code{length}. 

The total number of sites surveyed is \code{nrow(siteData)}. 
Duplicate site-level IDs are not allowed in \code{siteData}. 

See \bold{Input data frames} 
for when \code{detectionData} and \code{siteData} are required inputs.}

\item{area}{A scalar containing total area of 
inference. Commonly, this is study area size.  
If \code{area} is NULL (the default), 
area is set to 1 square unit of the output units. The default
produces abundance estimates that equal density estimates
(density estimates are always produced). 
If \code{area} is not NULL, it must have measurement units 
assigned by the \code{units} package. 
The units on \code{area} must be convertible
to squared output units. Units 
on \code{area} must be two-dimensional. 
For example, if output units are "foo", 
units on area must be convertible to "foo^2" by the \code{units}
package. 
Units of "km^2", "cm^2", "ha", "m^2", "acre", "mi^2", and many
others are acceptable.}

\item{ci}{A scalar indicating the confidence level of confidence intervals. 
Confidence intervals are computed using the bias corrected bootstrap
method. If \code{ci = NULL}, confidence intervals are not computed.}

\item{R}{The number of bootstrap iterations to conduct when \code{ci} is not
NULL.}

\item{plot.bs}{A logical scalar indicating whether to plot individual
bootstrap iterations.}

\item{bySite}{A logical scalar indicating whether to compute site-level
estimates of abundance. The default (\code{bySite=FALSE}) returns only one
overall abundance estimate. This routine does not calculate confidence
intervals for these site-level abundance estimates, so \code{ci} is set to
\code{NULL} if \code{bySite = TRUE}. See \code{\link{estimateN}}.}

\item{showProgress}{A logical indicating whether to show a text-based
progress bar during bootstrapping. Default is \code{TRUE}. 
It is handy to shut off the 
progress bar if running this within another function. Otherwise, 
it is handy to see progress of the bootstrap iterations.}

\item{control}{A list containing optimization control parameters such 
as the maximum number of iterations, tolerance, the optimizer to use, 
etc.  See the 
\code{\link{RdistanceControls}} function for explanation of each value,
the defaults, and the requirements for this list. 
See examples below for how to change controls.}
}
\value{
An 'abundance estimate' object, which is a list of
  class \code{c("abund", "dfunc")}, containing all the components of a "dfunc"
  object (see \code{dfuncEstim}), plus the following: 
  
  NOTE: THESE OUTPUT OBJECT DESCRIPTIONS NEED TO BE UPDATED. 
  
  \item{density}{Estimated density on the sampled area. The \emph{effectively}
  sampled area is 2*L*ESW (not 2*L*w.hi). Density has squared units of the 
  requested output units.  E.g., if \code{outputUnits} = "km" in the call 
  to \code{dfuncEstim}}, units on density are 1/km^2. No flexibility exists
  to have obtain different distance and density units (e.g., "m" and "ha", or "km" and 
  "acre").  To convert density to other units, use 
  \code{units::set_units(x$density, "<units>")}. 
  
 \item{abundance}{Estimated abundance in the study area (if \code{area} >
 1) or estimated density in the study area (if \code{area} = 1).}
 
  \item{n}{The number of detections
 (not individuals, unless all group sizes = 1) used in the estimate of
 abundance.}
 
  \item{area}{Total area of inference. Study area size}
  
  \item{effDistance}{Effective strip width for line-transects, effective
  radius for point-transects.  Both derived from \code{dfunc}. 
  See \code{\link{ESW}} or \code{\link{EDR}} for formulas.}
  
  \item{n.sites}{Total number of transects for line-transects, 
  total number of points for point-transects.}
  
  \item{tran.len}{Total transect length. NULL for point-transects.}
  
  \item{avg.group.size}{Average group size}
  
  \item{n.hat.ci}{The bias corrected bootstrap confidence interval for
  abundance (i.e., \code{n.hat}).  The names of this component 
  give the quantiles of the bootstrap distribution used during computation of
  the bias corrected interval.} 
  
  \item{density.ci}{The bias corrected bootstrap confidence interval for
  density (i.e., \code{density}).
  }

  \item{effDistance.ci}{The bias corrected bootstrap confidence interval for
  effective sampling distance (i.e., \code{effDistance}).
  }
  
  \item{B}{A data frame containing bootstrap density and effective distance 
  values.  Number of rows is always \code{R}, the number of bootstrap 
  requested iterations.  If a particular iteration did not converge, the
  corresponding row in \code{B} is all \code{NA} (e.g., use 'na.rm = TRUE' 
  when computing summaries). One column of the data frame 
  contains bootstrap values of \code{density}.  Another column contains bootstrap 
  values of \code{effDistance}. Bootstrap abundance values are 'B$density' * area.}
  
  \item{nItersConverged}{The number of bootstrap iterations that converged.  }
  
  \item{alpha}{The (scalar) confidence level of the
  confidence interval for \code{n.hat}.}
}
\description{
Estimate abundance (or density) given an estimated detection
  function and supplemental information on observed group sizes, transect
  lengths, area surveyed, etc.  Also computes confidence intervals of
  abundance (or density) using the bias corrected bootstrap method.
}
\details{
The abundance estimate for line transect surveys (if no covariates
   are included in the detection function) is 
   \deqn{N =\frac{n(A)}{2(ESW)(L)}}{%
         N = n*A / (2*ESW*L)} 
   where \emph{n} is total number of sighted individuals 
  (i.e., \code{sum(group.sizes)}), \emph{L} is the total length of 
  surveyed transect,
  and \emph{ESW} is effective strip width
  computed from the estimated distance function (i.e., \code{ESW(dfunc)}).
  
  The abundance estimate for point transect surveys (if no covariates are
  included) is 
   \deqn{N =\frac{n(A)}{\pi ESR^2 P}}{%
         N = n*A / ((3.1415)*ESR^2*(P))} 
   where \emph{n} is total number of sighted individuals,
   \emph{P} is the total number of surveyed points, 
   and \emph{ESR} is effective search radius 
   computed from the estimated distance function (i.e., \code{ESR(dfunc)}).

 Setting \code{plot.bs=FALSE} and \code{showProgress=FALSE} 
    suppresses all intermediate output.
}
\section{Bootstrap Confidence Intervals}{


  The bootstrap confidence interval for abundance 
  assumes that the fundamental units of
  replication (lines or points, hereafter "sites") are independent.
  The bias corrected bootstrap
  method used here resamples the units of replication (sites), 
  refits the distance function, and estimates abundance using 
  the resampled counts and re-estimated distance function. 
  The original data frames, \code{detectionData} and \code{siteData}, 
  are needed here for bootstrapping because they contain the transect 
  and detection information.
  If a double-observer data
  frame is included in \code{dfunc}, rows of the double-observer data frame
  are re-sampled each bootstrap iteration. 
  
  This routine does not 
  re-select the distance model fitted to resampled data.  The 
  model in the input object is re-fitted every iteration.  
  
  By default, \code{R} = 500 iterations are performed, after which the bias
  corrected confidence intervals are computed (Manly, 1997, section 3.4).
  
  During bootstrap iterations, the distance function can fail 
  to converge on the resampled data.   An iteration can fail 
  to converge for a two reasons:
  (1) no detections on the iteration, and (2) bad configuration 
  of distances on the iteration which pushes parameters to their 
  bounds. When an iteration fails to produce a valid 
  distance function, \code{Rdistance} 
  simply skips the intration, effectively ignoring these 
  non-convergent iterations. 
  If the proportion of non-convergent iterations is small 
  (less than 20% by default), the resulting confidence interval 
  on abundance is 
  probably valid.  If the proportion of non-convergent iterations 
  is not small (exceeds 20% by default), a warning is issued.  
  The print method (\code{print.abund}) is the routine that  issues this 
  warning. The warning can be 
  turned off by setting \code{maxBSFailPropForWarning} in the 
  print method to 1.0, or by modifying the code in \code{RdistanceControls()}
  to re-set the default threshold and storing the modified 
  function in your \code{.GlobalEnv}.  Additional iterations may be needed 
  to achieve an adequate number. Check number of convergent iterations by 
  counting non-NA rows in output data frame 'B'.
}

\examples{
# Load example sparrow data (line transect survey type)
data(sparrowDetectionData)
data(sparrowSiteData)

# Fit half-normal detection function
dfunc <- dfuncEstim(formula=dist~1
                    , detectionData=sparrowDetectionData
                    , likelihood="halfnorm"
                    , w.hi=units::set_units(100, "m")
                    , pointSurvey=FALSE)

# Estimate abundance given a detection function
# Note, a person should do more than R=20 iterations
fit <- abundEstim(dfunc
                , detectionData=sparrowDetectionData
                , siteData=sparrowSiteData
                , R=20
                , ci=0.95
                , area=units::set_units(10000, "m^2")
                , plot.bs=TRUE
                , bySite=FALSE)

# Print results
fit
         
}
\references{
Manly, B.F.J. (1997) \emph{Randomization, bootstrap, and 
  monte-carlo methods in biology}, London: Chapman and Hall.
  
  Buckland, S.T., D.R. Anderson, K.P. Burnham, J.L. Laake, D.L. Borchers,
   and L. Thomas. (2001) \emph{Introduction to distance sampling: estimating
   abundance of biological populations}. Oxford University Press, Oxford, UK.
}
\seealso{
\code{\link{dfuncEstim}}, \code{\link{autoDistSamp}}.
}
\author{
Trent McDonald, \email{trent@mcdonalddatasciences.com}\cr 
  Aidan McDonald, \email{aidan@mcdcentral.org}\cr 
  Jason Carlisle, Wyoming Game and Fish Department, 
  \email{jason.carlisle@wyo.gov}
}
\keyword{model}
