% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoDistSamp.R
\name{autoDistSamp}
\alias{autoDistSamp}
\title{Automated classical distance analysis.}
\usage{
autoDistSamp(formula, detectionData, siteData, w.lo = 0, w.hi = max(dist),
  likelihoods = c("halfnorm", "hazrate", "uniform", "negexp", "Gamma"),
  series = c("cosine", "hermite", "simple"), expansions = 0:3,
  pointSurvey = FALSE, warn = TRUE, area = 1, ci = 0.95, R = 500,
  bySite = FALSE, plot.bs = FALSE, plot = TRUE, criterion = "AICc", ...)
}
\arguments{
\item{formula}{This parameter is passed to \code{dfuncEstim}.
See \code{dfuncEstim} documentation for definition.}

\item{detectionData}{This parameter is passed to \code{dfuncEstim} 
and \code{abundEstim}. See \code{abundEstim} documentation for definition.}

\item{siteData}{This parameter is passed to \code{abundEstim}.
See \code{abundEstim} documentation for definition.}

\item{w.lo}{This parameter is passed to \code{dfuncEstim}.
See \code{dfuncEstim} documentation for definition.}

\item{w.hi}{This parameter is passed to \code{dfuncEstim}.
See \code{dfuncEstim} documentation for definition.}

\item{likelihoods}{Vector of strings specifying the 
likelihoods to consider during model selection. Valid values 
at present are "uniform", "halfnorm", "hazrate", "negexp", 
and "Gamma". See Details for the models this routine considers.}

\item{series}{Vector of series types to consider during model selection. 
Valid values are 'simple', 'hermite', and 'cosine'. See Details for 
the models this routine considers.}

\item{expansions}{Vector of the number of expansion terms to 
consider during model selection. Valid values are 0 through 3. 
See Details for the models this routine considers. Note, expansion terms are not currently
allowed in models with covariates.}

\item{pointSurvey}{This parameter is passed to \code{dfuncEstim}.
See \code{dfuncEstim} documentation for definition.}

\item{warn}{This parameter is passed to \code{dfuncEstim}.
\code{dfuncEstim} documentation for definition.}

\item{area}{This parameter is passed to \code{abundEstim}.
See \code{abundEstim} documentation for definition.}

\item{ci}{This parameter is passed to \code{abundEstim}.
See \code{abundEstim} documentation for definition.}

\item{R}{This parameter is passed to \code{abundEstim}.
See \code{abundEstim} documentation for definition.}

\item{bySite}{This parameter is passed to \code{abundEstim}.
See \code{abundEstim} documentation for definition.}

\item{plot.bs}{This parameter is passed to \code{abundEstim}.
See \code{abundEstim} documentation for definition.}

\item{plot}{Logical scalar specifying whether to plot models during model selection. 
If \code{TRUE}, a histogram with fitted distance function is plotted for every fitted model. 
The function pauses between each plot and prompts the user for whether they want to continue or not. 
For completely automated estimation, set \code{plot} = \code{FALSE}.}

\item{criterion}{A string specifying the criterion to use when assessing model fit.
The best fitting model from this routine is the one with lowest value 
of this fit criterion.  This must be one of "AICc" (the default), 
"AIC", or "BIC".  See \code{\link{AIC.dfunc}} for formulas.}

\item{...}{Additional parameters passed to \code{dfuncEstim}, which in turn are passed to \code{F.gx.estim}. 
These include \code{x.scl}, \code{g.x.scl}, and \code{observer} for estimating double observer probabilities.}
}
\value{
An 'abundance estimate' object (see \code{abundEstim} and \code{dfuncEstim}). 
  Returned abundance estimates are based on the best fitting distance function among those fitted.
}
\description{
Perform automated classical detection function selection and 
estimation of abundance.
}
\details{
During model selection, each series and number of expansions is crossed with 
each of the likelihoods. For example, if \code{likelihoods} has 3 elements, 
\code{series} has 2 elements, and \code{expansions} has 4 elements, 
the total number of models fitted is 3 (likelihoods) * 2 (series) * 4 (expansions) 
= 24 models.  The default specification fits 41 detection functions 
from the "halfnorm", "hazrate", "uniform", "negexp", and "Gamma" likelihoods 
(note that Gamma does not currently implement expansions, see 
\code{\link{Gamma.like}}). Note, expansion terms are not currently
allowed in models with covariates.  The model with lowest AIC is selected 
as 'best', and estimation of abundance proceeds using that model.
}
\examples{
# Load the example datasets of sparrow detections and transects from package
  data(sparrow.detections)
  data(sparrow.transects)
  
  # Automate fitting multiple detection functions
  # And estimate abundance (density per ha in this case) given the 'best' detection function
  # Note, area=10000 converts to density per hectare (for distances measured in meters)
  # Note, a person should do more than R=20 iterations 
  autoDistSamp(detectionData=sparrow.detections, siteData=sparrow.sites,
                  likelihood=c("halfnorm", "hazrate", "negexp"),
                  series=c("cosine", "simple"),
                  expansions=c(0, 1), area=10000, R=20, ci=0.95, bySite=FALSE,
                  plot.bs=FALSE, w.hi=150, plot=TRUE)
}
\seealso{
\code{\link{dfuncEstim}}, \code{\link{abundEstim}}
}
\author{
Trent McDonald, WEST Inc.,  \email{tmcdonald@west-inc.com}
        Aidan McDonald, WEST Inc.,  \email{aidan@mcdcentral.org}
        Jason Carlisle, University of Wyoming and WEST Inc., \email{jcarlisle@west-inc.com}
}
\keyword{model}
