% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Perform non-parallel bootstrap iterations}
\usage{
bootstrap(
  object,
  area,
  propUnitSurveyed,
  R,
  plot.bs,
  plotCovValues,
  showProgress,
  parallel,
  cores
)
}
\description{
Perform non-parallel bootstrap iterations
}
