% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLineLength.R
\name{calcLineLength}
\alias{calcLineLength}
\title{Calculate target transect length}
\usage{
calcLineLength(sPoly, N, p, w, targetGroups, avgGroupSize)
}
\arguments{
\item{sPoly}{An \code{sf} or \code{sfc} object containing the polygon(s)
(\code{MULTIPOLYGON} or \code{POLYGON} geometries) within which
transects will be placed. \code{sPoly} must be projected to a planar
coordinate system whose linear unit is meters (ideally an equal-area
projection so areas are undistorted). Geographic coordinates are rejected.}

\item{N}{Scalar, expected abundance of individuals in the study area.  Estimates
from past surveys are likely your best option here.  Used to calculate the
expected group density in the survey area.}

\item{p}{Scalar, expected probability of detecting a group.  If expected
strip width (ESW) is available, set \code{p} = \code{ESW/w}.}

\item{w}{Scalar, nominal width of the survey strip.  Must have measurement
units attached.}

\item{targetGroups}{Scalar, the desired number of detecterd groups
at the end of surveys.}

\item{avgGroupSize}{Scalar, expected average number of individuals in each
group.}
}
\value{
Scalar, the total length of transects to survey. Units of measurement
attached to the return are the square root of the polygon's area.
E.g., if \code{sPoly} is projected to a
standard UTM plane, their area is m^2, and transect length units will
be m.
}
\description{
Calculate the total length of transects to survey to reach a user-provided
target number of detected groups.  All else being equal, areas
with higher group densities will require less survey effort than
those with higher densities.
}
\examples{

data(exampleSurveyPoly)
calcLineLength(sPoly = exampleSurveyPoly[1, ]
               , N = 12000
               , p = 0.5
               , w = units::set_units(200, "m")
               , targetGroups = 300
               , avgGroupSize = 2.5
               )

}
\author{
Jason Carlisle, based on some clever back-of-the-napkin math by Trent
McDonald and Greg Hiatt at a WGFD training on pronghorn LT surveys in Laramie
in April 2022.
}
