% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNEvalPts.R
\name{checkNEvalPts}
\alias{checkNEvalPts}
\title{checkNEvalPts - Check number of numeric integration intervals}
\usage{
checkNEvalPts(nEvalPts)
}
\arguments{
\item{nEvalPts}{An integer to check.}
}
\value{
The first element of nEvalPts is returned if it is acceptable.
If nEvalPts is not acceptable, an error is thrown.
}
\description{
Check that number of integration intervals is odd and 
sufficiently large.
}
