% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfuncEstimErrMessage.R
\name{dfuncEstimErrMessage}
\alias{dfuncEstimErrMessage}
\title{dfuncEstimErrMessage - dfuncEstim error messages}
\usage{
dfuncEstimErrMessage(txt, attri)
}
\arguments{
\item{txt}{A text string describing the error.}

\item{attri}{An attribute to report.}
}
\value{
A string
}
\description{
Utility function to produce error messages suitable for \code{stop}
}
