% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiableLikelihoods.R
\name{differentiableLikelihoods}
\alias{differentiableLikelihoods}
\title{Differentiable likelihoods in Rdistance}
\usage{
differentiableLikelihoods()
}
\value{
A character vector of differentiable
likelihoods.
}
\description{
Return a character vector of
likelihoods which are differentiable and hence
the second derivative method can be used to estimate
variance-covariance.  Any likelihoods not in this list
must use bootstrapping. This vector is polled in
a few Rdistance routines, notably \code{parseModel}
and \code{varcovarEstim}.
}
