% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDR.R
\name{EDR}
\alias{EDR}
\title{Effective Detection Radius (EDR) for estimated detection functions with point transects}
\usage{
EDR(obj)
}
\arguments{
\item{obj}{An estimated detection function object.  An estimated detection
function object has class 'dfunc', and is usually produced by a call to 
\code{F.dfunc.estim}. The estimated detection function may optionally contain 
a \eqn{g(0)} component.  If no \eqn{g(0)} component is found, \eqn{g(0)} = 1
is assumed.}
}
\value{
A scalar equal to the area under the detection function from \code{obj$w.lo} to \code{obj$w.hi}.
}
\description{
Computes Effective Detection Radius (EDR) for estimated
detection functions with point transects.  The point-transect equivalent to
Effective Strip Width (ESW).
}
\details{
The point-transect equivalent to
Effective Strip Width (ESW).
}
\seealso{
\code{\link{F.dfunc.estim}} \code{\link{ESW}}
}
\author{
Aidan McDonald, WEST Inc., \email{aidan@mcdcentral.org}
}
\keyword{modeling}
