% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveDistance.R
\name{effectiveDistance}
\alias{effectiveDistance}
\title{effectiveDistance - Effective sampling distances}
\usage{
effectiveDistance(x, newdata = NULL)
}
\arguments{
\item{x}{An estimated detection function object, normally 
produced by calling \code{\link{dfuncEstim}}.}

\item{newdata}{A data frame containing new values for 
covariates at which either
ESW's or EDR's will be computed. If NULL and 
  \code{x} contains covariates, the  
  covariates stored in
  \code{x} are used (like \code{\link{predict.lm}}).
  If not NULL, covariate values in \code{newdata}
  are used. 
  See \bold{Value} section for more information.}
}
\value{
If \code{newdata} is present, the returned value is 
a vector of effective sampling distances for values of the 
covariates in \code{newdata} with length equal to 
the number of rows in \code{newdata}. 
If \code{newdata} is NULL, the returned value is a vector of effective
sampling distances for covariate values in \code{x} and has 
the number of detections in \code{x}.  The returned vector 
has measurement units, i.e., \code{x$outputUnits}.
}
\description{
Computes Effective Strip Width (ESW) for line-transect detection
  functions, or the analogous Effective Detection Radius (EDR) for point-transect
  detection functions.
}
\details{
Serves as a wrapper for 
\code{\link{ESW}} and \code{\link{EDR}}.
}
\seealso{
\code{\link{dfuncEstim}} \code{\link{ESW}} \code{\link{EDR}}
}
\keyword{modeling}
