% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveDistance.R
\name{effectiveDistance}
\alias{effectiveDistance}
\title{Calculates the ESW or EDR for estimated detection functions}
\usage{
effectiveDistance(obj, covars = NULL)
}
\arguments{
\item{obj}{An estimated detection function object.  An estimated detection
function object has class 'dfunc', and is usually produced by a call to 
\code{dfuncEstim}. The estimated detection function may optionally contain 
a \eqn{g(0)} component.  If no \eqn{g(0)} component is found, \eqn{g(0)} = 1
is assumed.}

\item{covars}{Covariate values for which to calculate ESW.}
}
\value{
A scalar equal to the area under the detection function from \code{obj$w.lo} to \code{obj$w.hi}.
}
\description{
Computes Effective Strip Width (ESW) for line-transect detection
functions, or the analagous Effective Detection Radius (EDR) point-transect
detection functions.
}
\details{
Serves as a wrapper for \code{\link{ESW}} and \code{\link{EDR}}.
}
\seealso{
\code{\link{dfuncEstim}} \code{\link{ESW}} \code{\link{EDR}}
}
\keyword{modeling}
