% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errDataUnk.R
\name{errDataUnk}
\alias{errDataUnk}
\title{Unknown error message}
\usage{
errDataUnk(txt, attri)
}
\arguments{
\item{txt}{Text.  The "unknown" we are looking for.}

\item{attri}{Attribute description we are looking for.}
}
\value{
A descriptive string, suitable for warning or error.
}
\description{
Constructs a string stating what is "unknown" that is
suitable for use in warning and error functions.
}
