% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateNhat.r
\name{estimateNhat}
\alias{estimateNhat}
\title{estimateNhat - Estimate abundance}
\usage{
estimateNhat(dfunc, detection.data, site.data, area)
}
\arguments{
\item{dfunc}{An estimate distance function (see \code{F.dfunc.estim}).}

\item{detection.data}{A data frame containing information on 
detections, such as distance and which transect or point they occured
on (see \code{F.dfunc.estim}).}

\item{site.data}{A data frame containing information on the 
transects or points surveyed  (see \code{F.dfunc.estim}).}

\item{area}{Total area of inference. Study area size.}
}
\value{
A list containing the following components:
   \item{dfunc}{The input distance function}
   \item{n.hat}{The estimate of abundance}
   \item{n}{Number of groups detected}
   \item{area}{Total area of inference. Study area size}
   \item{esw}{Effective strip width for line-transects, effective
   radius for point-transects.  Both derived from \code{dfunc}}
   \item{n.sites}{Total number of transects for line-transects, 
   total number of points for point-transects.}
   \item{tran.len}{Total transect length. NULL for point-transects.}
   \item{avg.group.size}{Average group size}
}
\description{
Estimate abundance given a distance function, 
detection data, site data, and area.  This is called internally 
by F.abund.estim during bootstrapping.
}
\examples{
# Load the example datasets of sparrow detections and transects from package
  data(sparrow.detections)
  data(sparrow.transects)
  
  # Fit detection function to perpendicular, off-transect distances
  dfunc <- F.dfunc.estim(sparrow.detections, w.hi=150)
  
  # Estimate abundance given a detection function
  n.hat <- estimate.nhat(dfunc, sparrow.detections, sparrow.sites)
  
}
\seealso{
\code{\link{F.dfunc.estim}}, \code{\link{F.abund.estim}}
}
\author{
Trent McDonald, WEST Inc.,  \email{tmcdonald@west-inc.com}\cr
        Jason Carlisle, University of Wyoming and WEST Inc, \email{jcarlisle@west-inc.com}\cr
        Aidan McDonald, WEST Inc., \email{aidan@mcdcentral.org}
}
\keyword{model}
