% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.maximize.g.r
\name{F.maximize.g}
\alias{F.maximize.g}
\title{F.maximize.g - Fincds the coordinate of the maximum of a distance function.}
\usage{
F.maximize.g(fit, covars = NULL)
}
\arguments{
\item{covars}{Covariate values to calculate maximum for.}

\item{x}{An estimated 'dfunc' object produced by \code{F.dfunc.estim}.}
}
\value{
The value of x that maximizes g(x) in \code{fit}.
}
\description{
Find the x coordinate that maximizes g(x).
}
\examples{
\dontrun{
# Fake data
set.seed(22223333)
x <- rgamma(100, 10, 1)

fit <- F.dfunc.estim( x, likelihood="Gamma", x.scl="max" )

F.maximize.g( fit )  # should be near 10.
fit$x.scl            # same thing
}
}
\seealso{
\code{\link{F.dfunc.estim}}
}
\author{
Trent McDonald, WEST Inc.,  \email{tmcdonald@west-inc.com}
}
\keyword{model}
