% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaModes.R
\name{GammaModes}
\alias{GammaModes}
\title{Modes of the Gamma distribution}
\usage{
GammaModes(params)
}
\arguments{
\item{parms}{A matrix of gamma distribution parameters. First column 
is the scale parameter. Second column is the shape parameter.}
}
\value{
A vector of the location of the gamma modes.
}
\description{
Compute mode (i.e., maximum) of Rdistance's version of the gamma 
distribution.
}
