% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazrate.like.R
\name{hazrate.like}
\alias{hazrate.like}
\title{hazrate.like - Hazard rate likelihood}
\usage{
hazrate.like(a, dist, covars)
}
\arguments{
\item{a}{A vector of covariate coefficients and other likelihood 
parameter values. Length 
should be (number of covariates including intercept) + 
(number of expansions) + (likelihood %in% c("hazrate","logistic"). 
Coefficients should all be on a log scale
because the link function for all likelihoods is exponential.}

\item{dist}{A numeric vector containing 
the observed detection distances.}

\item{covars}{A data frame or matrix containing values 
of the covariates for each distance observation.}
}
\value{
A list containing two components. One component, 
named \code{L.unscaled}, is a vector of the likelihood 
values at every
observed distance. L is "unscaled" because it does 
not integrate to one. Scaling happens elsewhere.

The second component, named \code{params},
is a data frame of the likelihood's (canonical) parameters 
for every input distance. Columns in the data frame
are named \emph{par1}, \emph{par2}, ..., \emph{parN}. Size
of the data frame is number of observations by number 
of canonical parameters.

Assuming \code{L} is the list returned by this function, 
the negative log likelihood is \code{-sum(log(L$L.unscaled / I), na.rm=T)}, 
where \code{I} is the integration constant, or 
area under the likelihood between 
\code{w.lo} and \code{w.hi}. 
Note that returned likelihood values for distances less 
than \code{w.lo} or greater than \code{w.hi} are \code{NA}; 
hence, \code{na.rm=TRUE} in the sum. 
Values in \code{L$L.unscaled} are always greater 
than or equal to zero.
}
\description{
Computes the hazard rate distance function.
}
\details{
The hazard rate likelihood is 
\deqn{f(x|\sigma,k) = 1 - \exp(-(x/\sigma)^{-k})}{%
f(x|Sigma,k) = 1 - exp(-(x/Sigma)^(-k))} 
where \eqn{\sigma}{Sigma} determines location 
(i.e., distance at which the function equals 1 - exp(-1) = 0.632), 
and \eqn{k}{k} determines slope of the function 
at \eqn{\sigma}{Sigma} (i.e., larger k equals steeper 
slope at \eqn{\sigma}{Sigma}). For distance analysis, 
the valid range for both \eqn{\sigma}{Sigma} and k is
\eqn{\geq 0}{>=0}.
}
\examples{
d <- seq(0, 100, length=100)
covs <- matrix(1,length(d),1)
hazrate.like(c(log(20), 5), d, covs)

# Changing location parameter
plot(d, hazrate.like(c(log(20), 5), d, covs)$L.unscaled, type="l", col="red")
lines(d, hazrate.like(c(log(40), 5), d, covs)$L.unscaled, col="blue")
abline(h = 1 - exp(-1), lty = 2)
abline(v = c(20,40), lty = 2)

# Changing slope parameter
plot(d, hazrate.like(c(log(50), 20), d, covs)$L.unscaled, type="l", col="red")
lines(d, hazrate.like(c(log(50), 2), d, covs)$L.unscaled, col="blue")
abline(h = 1 - exp(-1), lty = 2)
abline(v = 50, lty = 2)

         
}
\seealso{
\code{\link{dfuncEstim}},
         \code{\link{hazrate.like}},
         \code{\link{uniform.like}},
         \code{\link{negexp.like}},
         \code{\link{Gamma.like}},
         \code{\link{triangle.like}}
}
\keyword{models}
