% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateHalfnormLines.R
\name{integrateHalfnormLines}
\alias{integrateHalfnormLines}
\title{Integrate Half-normal line surveys}
\usage{
integrateHalfnormLines(
  object,
  newdata = NULL,
  w.lo = NULL,
  w.hi = NULL,
  Units = NULL
)
}
\arguments{
\item{object}{An Rdistance model frame or fitted distance function,
normally produced by a call to \code{\link{dfuncEstim}}.}

\item{newdata}{A data frame containing new values for 
covariates at which either
ESW's or EDR's will be computed. If NULL and 
  \code{object} contains covariates, the  
  covariates stored in
  \code{object} are used (like \code{\link{predict.lm}}).
  If not NULL, covariate values in \code{newdata}
  are used. 
  See \bold{Value} section for more information.}
}
\value{
A vector of areas under distance functions. 
If \code{newdata} is specified, return length is 
\code{nrow(newdata)}.  If \code{newdata} is NULL, 
return length is \code{length(distances(object))}.
}
\description{
Compute integral of the half-normal distance function for 
line-transect surveys.
}
\details{
Returned integral is computed using R's base 
function \code{pnorm()}, which is very accurate.
}
\examples{

# Fake distance function object w/ minimum inputs for integration
d <- units::set_units(rep(1,4),"m") # Only units needed, not values
obs <- factor(rep(c("obs1", "obs2"), 2))
beta <- c(3.5, -0.5)
w.hi <- 125
w.lo <- 20
ml <- list(
    mf = model.frame(d ~ obs)
  , par = beta 
  , likelihood = "halfnorm"
  , w.lo = units::set_units(w.lo, "m")
  , w.hi = units::set_units(w.hi, "m")
)
class(ml) <- "dfunc"
integrateHalfnormLines(ml)

# Check: Integral of exp(-x^2/(2*s^2)) from 0 to w.hi-w.lo
b <- exp(c(beta[1], beta[1] + beta[2]))
intgral <- (pnorm(w.hi, mean=w.lo, sd = b) - 0.5) * sqrt(2*pi)*b
intgral


}
\seealso{
\code{\link{integrateNumeric}}; \code{\link{integrateNegexp}}; 
\code{\link{integrateOneStepLines}}
}
