% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateHalfnormPoints.R
\name{integrateHalfnormPoints}
\alias{integrateHalfnormPoints}
\title{Integrate Half-normal Point transects}
\usage{
integrateHalfnormPoints(
  object,
  newdata = NULL,
  w.lo = NULL,
  w.hi = NULL,
  Units = NULL
)
}
\arguments{
\item{object}{An Rdistance model frame or fitted distance function,
normally produced by a call to \code{\link{dfuncEstim}}.}

\item{newdata}{A data frame containing new values for 
covariates at which either
ESW's or EDR's will be computed. If NULL and 
  \code{object} contains covariates, the  
  covariates stored in
  \code{object} are used (like \code{\link{predict.lm}}).
  If not NULL, covariate values in \code{newdata}
  are used. 
  See \bold{Value} section for more information.}
}
\value{
A vector of areas under distance functions. 
If \code{newdata} is specified, return length is 
\code{nrow(newdata)}.  If \code{newdata} is NULL, 
return length is \code{length(distances(object))}.
}
\description{
Compute integral of the half-normal distance function for 
point surveys.
}
\details{
Returned integral is 
\deqn{\int_0^{w.hi} xe^{-x^2/2\sigma^2} dx}{Integral(xe^{-x^2/2\sigma^2})}, 
which equals,
\deqn{0.5\sigma^2(1 - e^{-w^2/2\sigma^2})}{0.5*s^2*(1 - exp(-w^2/2*s^2))}.
}
\examples{

# Fake distance function object w/ minimum inputs for integration
d <- units::set_units(rep(1,4),"m") # Only units needed, not values
obs <- factor(rep(c("obs1", "obs2"), 2))
beta <- c(3.5, -0.5)
w.hi <- 125
w.lo <- 20
ml <- list(
    mf = model.frame(d ~ obs)
  , par = beta 
  , likelihood = "halfnorm"
  , w.lo = units::set_units(w.lo, "m")
  , w.hi = units::set_units(w.hi, "m")
)
class(ml) <- "dfunc"
integrateHalfnormPoints(ml)

# Check: Integral of x exp(-x^2/(2*s^2)) from 0 to w = w.hi-w.lo
sigma <- exp(c(beta[1], beta[1] + beta[2]))
w <- w.hi - w.lo
intgral <- sigma^2 * (1 - exp(-w^2 / (2*sigma^2)))
intgral

# Effective detection radius
sqrt(2 * intgral)

}
\seealso{
\code{\link{integrateNumeric}}; \code{\link{integrateNegexp}}; 
\code{\link{integrateOneStepPoints}}
}
