% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateKey.R
\name{integrateKey}
\alias{integrateKey}
\title{Compute and print distance function integration}
\usage{
integrateKey(ml, key, f0, plot = FALSE)
}
\arguments{
\item{ml}{The fitted object or model list}

\item{key}{The scaled distance function.}

\item{f0}{The value of f(0) when integrating line transects. 
This should be the ESW for the case.}

\item{plot}{Logical scalar.  If TRUE, plot a diagnostic consisting
of the distance function and approximating points we use for quadrature.}
}
\value{
Nothing.  Prints information on integrals to the screen.
}
\description{
Check several integration characteristics, 
and report them to screen.  This was designed to print 
info when user asks for higher verbose level. The scaled
distance function should integrate to 1.0 every iteration of 
maximization, and this function checks that.
}
