% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateKey.R
\name{integrateKey}
\alias{integrateKey}
\title{Compute and print distance function integration}
\usage{
integrateKey(ml, key, likExpan, f0)
}
\arguments{
\item{ml}{The fitted object or model list}

\item{key}{The scaled distance function.}

\item{likExpan}{A string specifying the current estimation case.
Format of the string is <likelihood>_<num expansions>_<points?>. 
E.g., "halfnorm_0_FALSE" specifies halfnormal, no expansions, and 
line transect data.}

\item{f0}{The value of f(0) when integrating line transects. 
This should be the ESW for the case.}
}
\value{
Nothing.  Prints information on integrals to the screen.
}
\description{
Check several integration characteristics, 
and report them to screen.  This was designed to print 
info when user asks for higher verbose level. The scaled
distance function should integrate to 1.0 every iteration of 
maximization, and this function checks that.
}
