% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateOneStepLines.R
\name{integrateOneStepLines}
\alias{integrateOneStepLines}
\title{Integrate Line-transect One-step function}
\usage{
integrateOneStepLines(object, newdata = NULL, Units = NULL)
}
\arguments{
\item{object}{Either an Rdistance fitted distance function,
(an object that inherits from class "dfunc"; usually produced 
by a call to \code{\link{dfuncEstim}}), or a matrix of canonical 
distance function parameters (e.g., matrix(fit$par,1)). 
If a matrix, each row corresponds to a 
distance function and each column is a parameter. Argument 
\code{newdata} is ignored if \code{object} is a matrix.}

\item{newdata}{A data frame containing new values for 
covariates at which either
ESW's or EDR's will be computed. If NULL and 
  \code{object} contains covariates, the  
  covariates stored in
  \code{object} are used (like \code{\link{predict.lm}}).
  If not NULL, covariate values in \code{newdata}
  are used. 
  See \bold{Value} section for more information.}

\item{Units}{Physical units to apply to the first column of 
\code{object} when it is a matrix.  Ignored if \code{object}
is a fitted Rdistance distance function.}
}
\value{
A vector of areas under distance functions. 
If \code{object} is a distance function and 
\code{newdata} is specified, return length is 
\code{nrow(newdata)}.  If \code{object} is a distance function and 
\code{newdata} is NULL, 
return length is \code{length(distances(object))}. If 
\code{object} is a matrix of parameters, return length is 
\code{nrow(object)}.
}
\description{
Compute exact integral of the one-step distance function for line 
transects.
}
\details{
Returned integral is exact.
}
\examples{

# Check:
w.hi <- 125
w.lo <- 0
s1 <- 40
s2 <- exp(log(s1) + log(0.5))
obs1Scaler <- (pnorm(w.hi, mean=w.lo, sd = s1) - 0.5) * sqrt(2*pi)*s1
obs2Scaler <- (pnorm(w.hi, mean=w.lo, sd = s2) - 0.5) * sqrt(2*pi)*s2
c(obs1Scaler, obs2Scaler)

}
\seealso{
\code{\link{integrateNumeric}}; \code{\link{integrateNegexp}}; 
\code{\link{integrateHalfnorm}}
}
