% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateOneStepNumeric.R
\name{integrateOneStepNumeric}
\alias{integrateOneStepNumeric}
\title{Numeric Integration of One-step Function}
\usage{
integrateOneStepNumeric(
  object,
  newdata = NULL,
  w.lo = NULL,
  w.hi = NULL,
  Units = NULL,
  expansions = NULL,
  series = NULL,
  isPoints = NULL
)
}
\arguments{
\item{object}{Either an Rdistance fitted distance function,
(an object that inherits from class "dfunc"; usually produced 
by a call to \code{\link{dfuncEstim}}), or a matrix of canonical 
distance function parameters (e.g., matrix(fit$par,1)). 
If a matrix, each row corresponds to a 
distance function and each column is a parameter. Argument 
\code{newdata} is ignored if \code{object} is a matrix.}

\item{newdata}{A data frame containing new values for 
covariates at which either
ESW's or EDR's will be computed. If NULL and 
  \code{object} contains covariates, the  
  covariates stored in
  \code{object} are used (like \code{\link{predict.lm}}).
  If not NULL, covariate values in \code{newdata}
  are used. 
  See \bold{Value} section for more information.}

\item{w.hi}{Maximum sighting distance. Ignored if \code{object} 
is a fitted Rdistance distance function.}

\item{Units}{Physical units to apply to the first column of 
\code{object} when it is a matrix.  Ignored if \code{object}
is a fitted Rdistance distance function.}
}
\value{
A vector of areas under distance functions. 
If \code{object} is a distance function and 
\code{newdata} is specified, return length is 
\code{nrow(newdata)}.  If \code{object} is a distance function and 
\code{newdata} is NULL, 
return length is \code{length(distances(object))}. If 
\code{object} is a matrix of parameters, return length is 
\code{nrow(object)}.
}
\description{
Compute integral of the one-step distance function
using numeric integration. This function is only called for 
oneStep functions that contain expansion factors.
}
\details{
The \code{\link{oneStep.like}} function has an extremely large 
discontinuity in Theta. Accurate numeric integration requires special 
use of the Trapazoid rule. Rdistance's Simpson's rule routine 
(\code{\link{integrateNumeric}}) will not work for oneStep.
}
\examples{

fit <- dfuncEstim(thrasherDf, dist~1, likelihood = "oneStep")
integrateOneStepPoints(fit, newdata = data.frame(`(Intercept)`=1))
EDR(fit, newdata = data.frame(`(Intercept)`=1))

# Check: 
Theta <- exp(fit$par[1])
Theta <- units::set_units(Theta, "m")
p <- fit$par[2]
w.hi <- fit$w.hi
w.lo <- fit$w.lo
g.at0 <- w.lo
g.atT <- Theta
g.atTPlusFuzz <- (((1-p) * Theta) / ((w.hi - Theta) * p))*Theta
g.atWhi <- (((1-p) * Theta) / ((w.hi - Theta) * p))*w.hi
area.0.to.T <- (Theta - w.lo) * (g.atT - g.at0) / 2 # triangle; Theta^2/2
area.T.to.w <- (w.hi - Theta) * (g.atTPlusFuzz + g.atWhi) / 2 # trapazoid
area <- area.0.to.T + area.T.to.w
edr <- sqrt( 2*area )

}
\seealso{
\code{\link{integrateNumeric}}; 
\code{\link{integrateOneStepLines}}; \code{\link{integrateOneStepPoints}}
}
