% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateOneStepPoints.R
\name{integrateOneStepPoints}
\alias{integrateOneStepPoints}
\title{Integrate Point survey One-step function}
\usage{
integrateOneStepPoints(object, newdata = NULL)
}
\arguments{
\item{object}{An Rdistance model frame or fitted distance function,
normally produced by a call to \code{\link{dfuncEstim}}.}

\item{newdata}{A data frame containing new values for 
covariates at which either
ESW's or EDR's will be computed. If NULL and 
  \code{object} contains covariates, the  
  covariates stored in
  \code{object} are used (like \code{\link{predict.lm}}).
  If not NULL, covariate values in \code{newdata}
  are used. 
  See \bold{Value} section for more information.}
}
\value{
A vector of areas under distance functions. 
If \code{newdata} is specified, return length is 
\code{nrow(newdata)}.  If \code{newdata} is NULL, 
return length is \code{length(distances(object))}.
}
\description{
Compute integral of the one-step distance function
for point-surveys.
}
\details{
Returned integral is exact.
}
\examples{

# Check:
w.hi <- 125
w.lo <- 0
s1 <- 40
s2 <- exp(log(s1) + log(0.5))
obs1Scaler <- (pnorm(w.hi, mean=w.lo, sd = s1) - 0.5) * sqrt(2*pi)*s1
obs2Scaler <- (pnorm(w.hi, mean=w.lo, sd = s2) - 0.5) * sqrt(2*pi)*s2
c(obs1Scaler, obs2Scaler)

}
\seealso{
[integrateNumeric()]; [integrateNegexp()]; 
[integrateOneStep()]
}
