% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercept.only.R
\name{intercept.only}
\alias{intercept.only}
\title{intercept.only - Detect intercept-only distance function}
\usage{
intercept.only(x)
}
\arguments{
\item{x}{An Rdistance model frame or fitted distance function.}
}
\value{
TRUE if \code{x} contains an intercept-only.
FALSE if \code{x} contains at least one detection-level 
or transect-level covariate in the detection function.
}
\description{
Utility function to detect whether a distance function has 
covariates beyond the intercept. If the model contains an 
intercept-only, effective distance is constant across detections
and short-cuts can be implemented in code.
}
