% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.points.R
\name{is.points}
\alias{is.points}
\title{is.points - Tests for point surveys}
\usage{
is.points(x, ...)
}
\arguments{
\item{x}{An Rdistance model frame or fitted distance function.}
}
\value{
TRUE if the model frame or fitted distance function 
contains point surveys.  FALSE if the model frame or distance 
function contains line transect surveys.
}
\description{
Determines whether a distance function is for a point survey 
or line survey.
}
