% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.smoothed.R
\name{is.smoothed}
\alias{is.smoothed}
\title{Tests for smoothed distance functions}
\usage{
is.smoothed(object)
}
\arguments{
\item{object}{An Rdistance model frame or fitted distance function,
normally produced by a call to \code{\link[=dfuncEstim]{dfuncEstim()}}.}
}
\value{
TRUE if the model frame or fitted distance function
arises from a non-parametric density smoother. FALSE if the
model frame or distance
function is a parameterized function.
}
\description{
Determines whether a distance function is a non-parametric
smooth or classic parameterized function.
}
