% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likeParamNames.r
\name{likeParamNames}
\alias{likeParamNames}
\title{Likelihood parameter names}
\usage{
likeParamNames(like.form)
}
\arguments{
\item{like.form}{A text string naming the form of the likelihood. 
An error is thrown if the likelihood is unknown.}
}
\value{
A vector of parameter names for the likelihood
}
\description{
Returns names of the likelihood parameters. This is 
a helper function and is not
necessary for estimation.  It is nice to label some 
parameters with descriptive
names like "sigma" or "k", depending on the likelihood.
}
