% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likeParamNames.r
\name{likeParamNames}
\alias{likeParamNames}
\title{Likelhood parameter names}
\usage{
likeParamNames(like.form)
}
\arguments{
\item{like.form}{A text string nameing the form of the likelihood.}
}
\value{
A vector of parameter names for that likelihood
}
\description{
Returns names of the likelihood paramters. This is 
a helper function and not
necessary.  It is a nice to label some outputs in Rdistance with 
parameter names like "sigma" or "knee", depending on the likelihood.
}
\details{
For user defined functions, ensure that the 
user defined start-limits function named <likelihood>.start.limits
can be evaluated on a distance of 1, 0 expansions, low limit of 0 
high limit of 1, and that it returns the parameter names as 
the \code{$names} component of the result.
}
\author{
Trent McDonald
}
