% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nCovars.R
\name{nCovars}
\alias{nCovars}
\title{Number of covariates}
\usage{
nCovars(X)
}
\arguments{
\item{X}{The X matrix of covariates, or a vector.}
}
\value{
An integer scalar
}
\description{
Return number of covariates in a distance model
}
\details{
The reason this routine is needed is that sometimes
we pass one row of covariates to a likelihood function.
If so, it may come in as a normal vector, not a matrix.
If a normal vector, ncol(X) does not work.
}
