% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observationType.R
\name{observationType}
\alias{observationType}
\title{Type of observations}
\usage{
observationType(x)
}
\arguments{
\item{x}{Either an estimated distance function, output by
\code{dfuncEstim}, or an Rdistance nested data frame, output by
\code{RdistDf}.}
}
\value{
One of the following values:
"single", "1given2", "2given1", or "both". If observation type
has not been assigned,
return is NULL.
}
\description{
Return the type of observations (single or multiple observers)
represented in either a fitted distance function or Rdistance data frame.
}
\details{
This function is a simple helper function.  If \code{x} is an
estimated distance object, it polls the \code{obsType}
attribute of the object's Rdistance data frame. If \code{x} is an Rdistance nested data frame, it
polls the \code{obsType} attribute.
}
