% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.dfunc.R
\name{predict.dfunc}
\alias{predict.dfunc}
\title{Predict method for dfunc objects}
\usage{
\method{predict}{dfunc}(object, newdata, type = c("parameters", "inflation"),
  ...)
}
\arguments{
\item{object}{An estimated dfunc object.  See \code{\link{dfuncEstim}}.}

\item{newdata}{A data frame containing new values of 
the covariates at which predictions are to be computed.}

\item{type}{The type of predictions desired. Type = "parameters"
predicts parameters of the likelihood function.  Type = "inflation"
predicts the inflation factor for all observations.  Inflation 
factors use likelihood parameters to compute effective sampling distances
(ESW or ESR) and inverts them.}
}
\value{
A vector of predicted parameters or inflation factors.
}
\description{
Predict likelihood parameters or inflation 
factors for distance function objects.
}
\author{
Trent McDonald
}
