% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dfunc.r
\name{print.dfunc}
\alias{print.dfunc}
\title{Print a distance function object.}
\usage{
\method{print}{dfunc}(x, ...)
}
\arguments{
\item{x}{An estimated distance function resulting from a call to \code{F.dfunc.estim}.}

\item{\dots}{Included for compatability with other print methods.  Ignored here.}
}
\value{
The input value of \code{obj} is invisibly returned.
}
\description{
Print method for distance functions produced by F.dfunc.estim, that are of class \code{dfund}.
}
\details{
The call, coefficients of the distanced function, whether the estimation converged, 
the likelihood and expansion function, and other statistics are printed.  At the bottom
of the output, the following quantities are printed,
\itemize{  
  \item \samp{Strip} : The left (\code{w.lo}) and right (\code{w.hi}) truncation values.
  \item \samp{Effective strip width or detection radius} : ESW or EDR as computed by \code{effectiveDistance}.
  \item \samp{Scaling} : The horizontal and vertical coordinates used to scale the distance function. 
  Usually, the horizontal coordinate is 0 and the vertical coordinate is 1 (i.e., g(0) = 1).
  \item \samp{Log likelihood} : Value of the maximized log likelihood.
  \item \samp{AIC} : Value of AIC for the distance function.
}
The number of digits printed is controled by \code{options()$digits}.
}
\examples{
# Load the example datasets of sparrow detections and transects from package
data(sparrow.detections)
data(sparrow.transects)

# Fit detection function to perpendicular, off-transect distances
dfunc <- F.dfunc.estim(sparrow.detections, w.hi=150)

# Print the output                 
print(dfunc)
}
\seealso{
\code{\link{F.dfunc.estim}}, \code{\link{plot.dfunc}}, \code{\link{print.abund}}
}
\author{
Trent McDonald, WEST, Inc. \email{tmcdonald@west-inc.com}
        Aidan McDonald, WEST, Inc. \email{aidan@mcdcentral.org}
}
\keyword{models}
