% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pronghornAreas.R, R/pronghornDfDocumentation.R
\docType{data}
\name{pronghornAreas}
\alias{pronghornAreas}
\alias{pronghornDf}
\title{Pronghorn Line Transect Data}
\format{
A rowwise tibble containing 300 rows and 6 columns.  Each row
represents one transect.  Column \code{detections} contains an embedded data frame
of detections made on the transect of that row.

\strong{Transects}:
Transect columns are as follows:
\itemize{
\item \code{transectID}: (character) Unique transect identifier. This is
a concatenation of \code{herd}, \code{year}, and \code{lineLabel}.
\item \code{herd}: (character) Abbreviation of the surveyed herd unit. Herds are
considered distinct populations.
\item \code{year}: (character) Year of the survey.
\item \code{lineLabel}: (character) Unique identifier of the transect line
within herd and year.
\item \code{lineLength}: (numeric, kilometers) Length of the transect.
\item \code{detections}: (nested dataframe) Detections made on the transect
(next section).  NA if no detections were made on the transect.
}

\strong{Detections}:
The embedded data frame in column \code{detections} contains the following
variables:
\itemize{
\item \code{nominalDistBandLabel}: (character) The nominal distance bin identifier
assigned by observers in the aircraft at the time of detection based on sighting
a group between markers on the aircraft's wing strut.
\item \code{nominalDistBand}: (factor) The nominal distance bin endpoints assuming
an aircraft flight height of 300 [ft] above ground level (AGL) at the time of detection.
\item \code{flightHeight}: (numeric, feet) Height of the aircraft AGL at the time of detection.
\item \code{clusterSize}: (integer) Number of pronghorn in the detected group.
\item \code{adjustedMidpoint}: (numeric, meters) The distance bin's midpoint
after adjustment for AGL at the time of detection.
\item \code{adjustedBandLower}: (numeric, meters) The lower endpoint of the
distance bin after adjustment for AGL at the time of detection.
\item \code{adjustedBandUpper}: (numeric, meters) The upper endpoint of the
distance bin after adjustment for AGL at the time of detection.
}
}
\source{
These data arise from routine pronghorn population monitoring
conducted by the Wyoming Game and Fish Department.  Included with permission.
Contact: Dr. Jason Carlisle (Rdistance author).
}
\description{
Pronghorn (\emph{Antilocapra americana}) aerial line transect data collected
by the Wyoming Game and Fish Department in southeast Wyoming, 2012 - 2019.
Data contain transect and pronghorn group detection information
collected during five aerial surveys of four herd units during four years.
The general survey protocol is described in Johnson et al. (1991) and
Smyser et al. (2016), modified for these surveys to use 5 distance bins
(specified in Details below) and one observer surveying only one side of
the aircraft at a time.
}
\details{
\strong{Study Areas}:
Researchers placed transects systematically over polygons outlining habitable
areas within each herd unit's boundaries.
Habitable areas within each herd unit had the following sizes
(available in \code{data(pronghornAreas)}):\tabular{cr}{
   Herd \tab Area \cr
   --------- \tab -------------- \cr
   BA \tab 1004 [mi^2] \cr
   CE \tab 1420 [mi^2] \cr
   CO \tab 444 [mi^2] \cr
   MB \tab 3000 [mi^2] \cr
}


\strong{Nominal Distance Bins}:
Researchers established the following minimum and maximum distances for each
off-transect distance bin after multiple flights calibrating wing strut
markers against known-width ground objects (e.g., highways). These lower and
upper bin endpoints correspond to a flight altitude of 300 [ft] AGL.\tabular{crr}{
   Bin \tab Minimum \tab Maximum \cr
   --- \tab ----------- \tab ----------- \cr
   A \tab 65 [m] \tab 85 [m] \cr
   B \tab 85 [m] \tab 110 [m] \cr
   C \tab 110 [m] \tab 145 [m] \cr
   D \tab 145 [m] \tab 210 [m] \cr
   E \tab 210 [m] \tab 265 [m] \cr
}
}
\examples{
transectInfo <- pronghornDf |> 
  dplyr::reframe( herd = herd[1]
                , year = year[1]
                , lineLabel = lineLabel[1]
                , length = lineLength[1]
                , groups = ifelse(is.null(detections), 0, nrow(detections))
                , individuals = ifelse(is.null(detections), 0, sum(detections$clusterSize))
                )
surveyInfo <- transectInfo |> 
  dplyr::group_by(herd, year) |> 
  dplyr::summarise(transects = dplyr::n()
                 , transectTotLength = sum(length)
                 , totGroups = sum(groups)
                 , totIndividuals = sum(individuals)
                 , zeroTransects = sum(groups == 0)
  )
  
phDists <- unnest(pronghornDf) # no zero transects: okay for histo
hist(phDists$adjustedMidpoint
  , breaks = c(65,85,110,145,210,265,375)
  , prob = TRUE
  , xlab = "Offtransect distance")

}
\references{
Johnson, B. K., F. G. Lindzey, and R. J. Guenzel. 1991
Use of Aerial Line Transect Surveys to Estimate Pronghorn Populations in
Wyoming. \emph{Wildlife Society Bulletin} 19(3):315–321.
http://www.jstor.org/stable/3782522.

Smyser, T. J., R. J. Guenzel, C. N. Jacques, and E. O. Garton. 2016.
Double-observer evaluation of pronghorn aerial line-transect surveys.
\emph{Wildlife Research} 43(6):474–481.
}
