% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondDeriv.r
\name{secondDeriv}
\alias{secondDeriv}
\title{secondDeriv}
\usage{
secondDeriv(x, FUN, eps = 1e-08, ...)
}
\arguments{
\item{x}{The location (a vector) where the second derivatives 
of \code{FUN} are desired.}

\item{FUN}{An R function for which the second derivatives are 
sought.  
This must be a function of the form FUN <- function(x, ...){...}
where x is a vector of variable parameters to FUN at which 
to evaluate the 2nd derivative, 
and ... are additional paramters needed to evaluate the function.
FUN must return a single value (scalar), the height of the 
surface above \code{x}, i.e., FUN evaluated at x.}

\item{eps}{A vector of small distances to add to \code{x} 
when evaluating derivatives.   This is the 'delta x' or eps of 
the numerical derivatives. Default value 1e-8 for all 
dimensions.  If length of \code{eps} is less than length of \code{x},
\code{eps} is replicated to the length of \code{x}.}

\item{\dots}{Any arguments passed to \code{FUN}.}
}
\description{
Computes numeric 2nd derivatives (hessian) of an 
arbitrary multidimensional function at a particular location.
}
\details{
This function uses the "5-point" numeric second derivative 
method advocated in numerous numerical recipie texts.  During computation
of the 2nd derivative, FUN nust be 
capable of being evaluated at locations within a hyper-elipsoid 
with cardinal radii 2*\code{x}*(eps)^0.25, or 0.02*\code{x} at the 
default value for \code{eps}.   

A handy way to use this function is to call an optimization routine 
like \code{nlminb} with FUN, then call this function with the 
optimized values (solution) and FUN.  This will yeild the hessian 
at the solution rather than the hessian at the previous step of the 
optimization.
}
\examples{

func <- function(x){-x*x} # second derivative should be -2
secondDeriv(0,func)
secondDeriv(3,func)

func <- function(x){3 + 5*x^2 + 2*x^3} # second derivative should be 10+12x
secondDeriv(0,func)
secondDeriv(2,func)

func <- function(x){x[1]^2 + 5*x[2]^2} # should be rbind(c(2,0),c(0,10))
secondDeriv(c(1,1),func)

}
\author{
Trent McDonald
}
