\name{sparrows}
\docType{data}
\alias{sparrows}
\title{Example sparrow distance sampling data}
\description{
  This dataset demonstrates the type and format of input data required by \code{Rdistance} to estimate a detection function and abundance from distance sampling data collected by surveying line transects.  Data come from 72 transects, each 500 meters long, surveyed by the Wyoming Cooperative Fish & Wildlife Research Unit.
}
\usage{sparrows}
\format{A data.frame containing ___ observations.}
\source{Jason Carlisle, University of Wyoming, \email{jason.d.carlisle@gmail.com}}
\references{Carlisle and Chalfoun, unpublished data}
\keyword{datasets}