% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transectType.R
\name{transectType}
\alias{transectType}
\title{transectType - Type of transects}
\usage{
transectType(x)
}
\arguments{
\item{x}{Either an estimated distance function, output by 
\code{dfuncEstim}, or an Rdistance nested data frame, output by 
\code{RdistDf}.}
}
\value{
A scalar. Either 'line' if \code{x} contains  
continuous line-transect detections, or 'point' if \code{x} contains 
point-transects detections. If transect type has not been assigned, 
return is NULL.
}
\description{
Return the type of transects represented in either a fitted 
distance function or Rdistance data frame.
}
\details{
This function is a simple helper function.  If \code{x} is an
estimated distance object, it polls the \code{transType} attribute 
of \code{x}'s Rdistance nested data frame. 
If \code{x} is an Rdistance nested data frame, it 
polls the \code{transType} attribute.
}
