% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest}
\alias{unnest}
\title{unnest - Unnest an RdistDf data frame}
\usage{
unnest(x, ...)
}
\value{
An expanded data frame, without embedded data frames. Rows 
in the return represent with one detection or one transect.  If multiple
detections were made on one transect, the transect will appear on multiple
rows.  If no detections were made on a transect, it will appear on one 
row with NA detection distance.
}
\description{
Unnest an RdistDf data frame by expanding 
the embedded 'detections' column.  This unnest
includes the so-called zero transects (transects without detections).
}
\examples{
data('sparrowDf')

# tidyr::unnest() does not include zero transects
detectionDf <- tidyr::unnest(sparrowDf, detections)
nrow(detectionDf)
any(detectionDf$siteID == "B2")

# Rdistance::unnest() includes zero transects
fullDf <- unnest(sparrowDf)
nrow(fullDf)
any(fullDf$siteID == "B2")

}
